ten- add: BuyPoint update: Load/Save
